First apparently fully working version of the matchup code. Seems like it generated a good set of matchups between jaunty and karmic.
This code ended up being hairier than I was expecting for what in the end was a simple set of matchups. A much simpler option would just be to generate all the straightforward equal matchups and treat all others with manual exceptions (47 in this case). Would have saved lots of time to get to the same list.

Added tests for rule_enging and sources_info whenever a bug needed to be fixed. Test driven bug fixing I guess.